pyproject.toml
[tool.poetry]
name = "cue"
version = "0.1.0"
description = "Jira integration CLI"
authors = ["silkyanteater <cyclopesrufus@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
pyyaml = "^5.3.1"
toml = "^0.10.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
